+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/customization/MethodCustomization.java
@@ -3,65 +3,33 @@
 
 package com.microsoft.typespec.http.client.generator.core.customization;
 
-import com.microsoft.typespec.http.client.generator.core.customization.implementation.Utils;
-import com.microsoft.typespec.http.client.generator.core.customization.implementation.ls.EclipseLanguageClient;
 import java.lang.reflect.Modifier;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import org.eclipse.lsp4j.FileChangeType;
-import org.eclipse.lsp4j.FileEvent;
-import org.eclipse.lsp4j.Position;
-import org.eclipse.lsp4j.Range;
-import org.eclipse.lsp4j.SymbolInformation;
-import org.eclipse.lsp4j.TextEdit;
-import org.eclipse.lsp4j.WorkspaceEdit;
 
 /**
  * The method level customization for an AutoRest generated method.
  */
-public final class MethodCustomization extends CodeCustomization {
-    private final String packageName;
-    private final String className;
-    private final String methodName;
-    private final String methodSignature;
-
-    MethodCustomization(Editor editor, EclipseLanguageClient languageClient, String packageName, String className,
-        String methodName, String methodSignature, SymbolInformation symbol) {
-        super(editor, languageClient, symbol);
-        this.packageName = packageName;
-        this.className = className;
-        this.methodName = methodName;
-        this.methodSignature = methodSignature;
-    }
-
+public interface MethodCustomization extends CodeCustomization {
     /**
      * Gets the name of the method this customization is using.
      *
      * @return The name of the method.
      */
-    public String getMethodName() {
-        return methodName;
-    }
+    String getMethodName();
 
     /**
      * Gets the name of the class containing the method.
      *
      * @return The name of the class containing the method.
      */
-    public String getClassName() {
-        return className;
-    }
+    String getClassName();
 
     /**
      * Gets the Javadoc customization for this method.
      *
      * @return the Javadoc customization
      */
-    public JavadocCustomization getJavadoc() {
-        return new JavadocCustomization(editor, languageClient, fileUri, fileName,
-            symbol.getLocation().getRange().getStart().getLine());
-    }
+    JavadocCustomization<?> getJavadoc();
 
     /**
      * Rename a method in the class. This is a refactor operation. All references to this method across the client
@@ -70,12 +38,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param newName the new name for the method
      * @return the current method customization for chaining
      */
-    public MethodCustomization rename(String newName) {
-        WorkspaceEdit edit = languageClient.renameSymbol(fileUri, symbol.getLocation().getRange().getStart(), newName);
-        Utils.applyWorkspaceEdit(edit, editor, languageClient);
-
-        return refreshCustomization(methodSignature.replace(methodName + "(", newName + "("));
-    }
+    MethodCustomization rename(String newName);
 
     /**
      * Add an annotation to a method in the class.
@@ -83,9 +46,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param annotation the annotation to add. The leading @ can be omitted.
      * @return the current method customization for chaining
      */
-    public MethodCustomization addAnnotation(String annotation) {
-        return Utils.addAnnotation(annotation, this, () -> refreshCustomization(methodSignature));
-    }
+    MethodCustomization addAnnotation(String annotation);
 
     /**
      * Remove an annotation from the method.
@@ -93,16 +54,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param annotation the annotation to remove from the method. The leading @ can be omitted.
      * @return the current method customization for chaining
      */
-    public MethodCustomization removeAnnotation(String annotation) {
-        return Utils.removeAnnotation(this, compilationUnit -> compilationUnit.getClassByName(className)
-            .get()
-            .getMethodsByName(methodName)
-            .stream()
-            .filter(method -> Utils.declarationContainsSymbol(method.getRange().get(), symbol.getLocation().getRange()))
-            .findFirst()
-            .get()
-            .getAnnotationByName(Utils.cleanAnnotationName(annotation)), () -> refreshCustomization(methodSignature));
-    }
+    MethodCustomization removeAnnotation(String annotation);
 
     /**
      * Replace the modifier for this method.
@@ -117,12 +69,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @throws IllegalArgumentException If the {@code modifier} is less than to {@code 0} or any {@link Modifier}
      * included in the bitwise OR isn't a valid method {@link Modifier}.
      */
-    public MethodCustomization setModifier(int modifiers) {
-        Utils.replaceModifier(symbol, editor, languageClient, "(?:.+ )?(\\w+ )" + methodName + "\\(",
-            "$1" + methodName + "(", Modifier.methodModifiers(), modifiers);
-
-        return refreshCustomization(methodSignature);
-    }
+    MethodCustomization setModifier(int modifiers);
 
     /**
      * Replace the parameters of the method.
@@ -130,9 +77,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param newParameters New method parameters.
      * @return The updated MethodCustomization object.
      */
-    public MethodCustomization replaceParameters(String newParameters) {
-        return replaceParameters(newParameters, null);
-    }
+    MethodCustomization replaceParameters(String newParameters);
 
     /**
      * Replaces the parameters of the method and adds any additional imports required by the new parameters.
@@ -142,18 +87,7 @@ public final class MethodCustomization extends CodeCustomization {
      * are ambiguous on which to use such as {@code List} or the utility class {@code Arrays}.
      * @return A new MethodCustomization representing the updated method.
      */
-    public MethodCustomization replaceParameters(String newParameters, List<String> importsToAdd) {
-        String newSignature = methodName + "(" + newParameters + ")";
-
-        ClassCustomization classCustomization
-            = new PackageCustomization(editor, languageClient, packageName).getClass(className);
-
-        ClassCustomization updatedClassCustomization
-            = Utils.addImports(importsToAdd, classCustomization, classCustomization::refreshSymbol);
-
-        return Utils.replaceParameters(newParameters, updatedClassCustomization.getMethod(methodSignature),
-            () -> updatedClassCustomization.getMethod(newSignature));
-    }
+    MethodCustomization replaceParameters(String newParameters, List<String> importsToAdd);
 
     /**
      * Replace the body of the method.
@@ -161,9 +95,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param newBody New method body.
      * @return The updated MethodCustomization object.
      */
-    public MethodCustomization replaceBody(String newBody) {
-        return replaceBody(newBody, null);
-    }
+    MethodCustomization replaceBody(String newBody);
 
     /**
      * Replaces the body of the method and adds any additional imports required by the new body.
@@ -173,16 +105,7 @@ public final class MethodCustomization extends CodeCustomization {
      * are ambiguous on which to use such as {@code List} or the utility class {@code Arrays}.
      * @return A new MethodCustomization representing the updated method.
      */
-    public MethodCustomization replaceBody(String newBody, List<String> importsToAdd) {
-        ClassCustomization classCustomization
-            = new PackageCustomization(editor, languageClient, packageName).getClass(className);
-
-        ClassCustomization updatedClassCustomization
-            = Utils.addImports(importsToAdd, classCustomization, classCustomization::refreshSymbol);
-
-        return Utils.replaceBody(newBody, updatedClassCustomization.getMethod(methodSignature),
-            () -> updatedClassCustomization.getMethod(methodSignature));
-    }
+    MethodCustomization replaceBody(String newBody, List<String> importsToAdd);
 
     /**
      * Change the return type of the method. The new return type will be automatically imported.
@@ -198,9 +121,7 @@ public final class MethodCustomization extends CodeCustomization {
      * @param returnValueFormatter the return value String formatter as described above
      * @return the current method customization for chaining
      */
-    public MethodCustomization setReturnType(String newReturnType, String returnValueFormatter) {
-        return setReturnType(newReturnType, returnValueFormatter, false);
-    }
+    MethodCustomization setReturnType(String newReturnType, String returnValueFormatter);
 
     /**
      * Change the return type of the method. The new return type will be automatically imported.
@@ -219,93 +140,6 @@ public final class MethodCustomization extends CodeCustomization {
      * parentheses, opening and closing of code blocks have to be taken care of in the {@code returnValueFormatter}.
      * @return the current method customization for chaining
      */
-    public MethodCustomization setReturnType(String newReturnType, String returnValueFormatter,
-        boolean replaceReturnStatement) {
-        List<TextEdit> edits = new ArrayList<>();
-
-        int line = symbol.getLocation().getRange().getStart().getLine();
-        Position start = new Position(line, 0);
-        String oldLineContent = editor.getFileLine(fileName, line);
-        Position end = new Position(line, oldLineContent.length());
-        String newLineContent = oldLineContent.replaceFirst("(\\w.* )?(\\w+) " + methodName + "\\(",
-            "$1" + newReturnType + " " + methodName + "(");
-        TextEdit signatureEdit = new TextEdit();
-        signatureEdit.setNewText(newLineContent);
-        signatureEdit.setRange(new Range(start, end));
-        edits.add(signatureEdit);
-
-        String methodIndent = Utils.getIndent(editor.getFileLine(fileName, line));
-        String methodContentIndent = Utils.getIndent(editor.getFileLine(fileName, line + 1));
-        String oldReturnType = oldLineContent.replaceAll(" " + methodName + "\\(.*", "")
-            .replaceFirst(methodIndent + "(\\w.* )?", "")
-            .trim();
-        int returnLine = -1;
-        while (!oldLineContent.startsWith(methodIndent + "}")) {
-            if (oldLineContent.contains("return ")) {
-                returnLine = line;
-            }
-            oldLineContent = editor.getFileLine(fileName, ++line);
-        }
-        if (returnLine == -1) {
-            // no return statement, originally void return type
-            editor.insertBlankLine(fileName, line, false);
-            FileEvent blankLineEvent = new FileEvent();
-            blankLineEvent.setUri(fileUri);
-            blankLineEvent.setType(FileChangeType.Changed);
-            languageClient.notifyWatchedFilesChanged(Collections.singletonList(blankLineEvent));
-
-            TextEdit returnEdit = new TextEdit();
-            returnEdit.setRange(new Range(new Position(line, 0), new Position(line, 0)));
-            returnEdit.setNewText(methodContentIndent + "return " + returnValueFormatter + ";");
-            edits.add(returnEdit);
-        } else if (newReturnType.equals("void")) {
-            // remove return statement
-            TextEdit returnEdit = new TextEdit();
-            returnEdit.setNewText("");
-            returnEdit.setRange(new Range(new Position(returnLine, 0), new Position(line, 0)));
-            edits.add(returnEdit);
-        } else {
-            // replace return statement
-            TextEdit returnValueEdit = new TextEdit();
-            String returnLineText = editor.getFileLine(fileName, returnLine);
-            returnValueEdit
-                .setRange(new Range(new Position(returnLine, 0), new Position(returnLine, returnLineText.length())));
-            returnValueEdit.setNewText(returnLineText.replace("return ", oldReturnType + " returnValue = "));
-            edits.add(returnValueEdit);
-
-            editor.insertBlankLine(fileName, line, false);
-            FileEvent blankLineEvent = new FileEvent();
-            blankLineEvent.setUri(fileUri);
-            blankLineEvent.setType(FileChangeType.Changed);
-            languageClient.notifyWatchedFilesChanged(Collections.singletonList(blankLineEvent));
-
-            TextEdit returnEdit = new TextEdit();
-            returnEdit.setRange(new Range(new Position(line, 0), new Position(line, 0)));
-
-            if (replaceReturnStatement) {
-                returnEdit.setNewText(String.format(returnValueFormatter, "returnValue"));
-            } else {
-                returnEdit.setNewText(
-                    methodContentIndent + "return " + String.format(returnValueFormatter, "returnValue") + ";");
-            }
-
-            edits.add(returnEdit);
-        }
-
-        WorkspaceEdit workspaceEdit = new WorkspaceEdit();
-        workspaceEdit.setChanges(Collections.singletonMap(fileUri, edits));
-        Utils.applyWorkspaceEdit(workspaceEdit, editor, languageClient);
-
-        Utils.organizeImportsOnRange(languageClient, editor, fileUri, new Range(start, end));
-
-        String newMethodSignature
-            = methodSignature.replace(oldReturnType + " " + methodName, newReturnType + " " + methodName);
-
-        return refreshCustomization(newMethodSignature);
-    }
-
-    private MethodCustomization refreshCustomization(String methodSignature) {
-        return new PackageCustomization(editor, languageClient, packageName).getClass(className)
-            .getMethod(methodSignature);
-    }
+    MethodCustomization setReturnType(String newReturnType, String returnValueFormatter,
+        boolean replaceReturnStatement);
 }
