+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/extension/plugin/JavaSettings.java
@@ -6,9 +6,11 @@ package com.microsoft.typespec.http.client.generator.core.extension.plugin;
 import com.azure.json.JsonProviders;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonToken;
+import com.github.javaparser.JavaParser;
 import com.microsoft.typespec.http.client.generator.core.mapper.Mappers;
 import com.microsoft.typespec.http.client.generator.core.mapper.azurevnext.AzureVNextMapperFactory;
 import com.microsoft.typespec.http.client.generator.core.mapper.clientcore.ClientCoreMapperFactory;
+import com.microsoft.typespec.http.client.generator.core.postprocessor.Postprocessor;
 import com.microsoft.typespec.http.client.generator.core.template.Templates;
 import com.microsoft.typespec.http.client.generator.core.template.azurevnext.AzureVNextTemplateFactory;
 import com.microsoft.typespec.http.client.generator.core.template.clientcore.ClientCoreTemplateFactory;
@@ -385,7 +387,10 @@ public class JavaSettings {
         this.shareJsonSerializableCode = getBooleanValue(host, "share-jsonserializable-code", false);
 
         // Whether to use object for unknown.
-        this.useObjectForUnknown = getBooleanValue(host, "use-object-for-unknown", false);
+        this.useObjectForUnknown = getBooleanValue(host, "use-object-for-unknown", true);
+
+        // Whether to use Eclipse Language Server when running code customizations.
+        this.useEclipseLanguageServer = getBooleanValue(host, "use-eclipse-language-server", false);
     }
 
     private void updateFlavorFactories() {
@@ -1503,6 +1508,24 @@ public class JavaSettings {
 
     public boolean isUseObjectForUnknown() {
         return useObjectForUnknown;
+    }
+
+    private final boolean useEclipseLanguageServer;
+
+    /**
+     * If there is {@link Postprocessor} code customizations to run, this determines whether to use the Eclipse Language
+     * Server to run the code customizations.
+     * <p>
+     * This is a temporary setting until the Eclipse Language Server is removed and code customizations use
+     * {@link JavaParser} to run customizations. Switching to {@link JavaParser} will make customizations run faster but
+     * be more complicated to implement, where the latter part is okay as they're meant to be a "break the glass"
+     * situation where Swagger / TypeSpec definitions cannot convey what is required in code. Until this feature flag
+     * was added, code customizations were commonly used to gloss over issues with the Swagger / TypeSpec definitions.
+     *
+     * @return Whether to use the Eclipse Language Server to run the code customizations.
+     */
+    public boolean isUseEclipseLanguageServer() {
+        return useEclipseLanguageServer;
     }
 
     private static final String DEFAULT_CODE_GENERATION_HEADER
