+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/postprocessor/Postprocessor.java
@@ -28,10 +28,12 @@ import org.slf4j.Logger;
 
 public class Postprocessor {
     protected final NewPlugin plugin;
+    private final boolean useEclipseLanguageServer;
     private final Logger logger;
 
-    public Postprocessor(NewPlugin plugin) {
+    public Postprocessor(NewPlugin plugin, boolean useEclipseLanguageServer) {
         this.plugin = plugin;
+        this.useEclipseLanguageServer = useEclipseLanguageServer;
         this.logger = new PluginLogger(plugin, Postprocessor.class);
     }
 
@@ -98,7 +100,7 @@ public class Postprocessor {
             try {
                 Customization customization = customizationClass.getConstructor().newInstance();
                 logger.info("Running customization, this may take a while...");
-                fileContents = customization.run(fileContents, logger);
+                fileContents = customization.run(fileContents, useEclipseLanguageServer, logger);
             } catch (Exception e) {
                 logger.error("Unable to complete customization", e);
                 throw new RuntimeException("Unable to complete customization", e);
