+++ packages/http-client-java/generator/http-client-generator-test/Generate.ps1
@@ -90,6 +90,13 @@ $generateScript = {
     $tspOptions += " --option ""@typespec/http-client-java.customization-class=../../customization/src/main/java/CustomizationTest.java"""
   }
 
+  # Test customization using only JavaParser for one of the TypeSpec definitions - naming-javaparser.tsp
+  if ($tspFile -match "tsp[\\/]naming-javaparser.tsp$") {
+      # Add the customization-class option for Java emitter
+      $tspOptions += " --option ""@typespec/http-client-java.customization-class=../../customization/src/main/java/JavaParserCustomizationTest.java"""
+      $tspOptions += " --option ""@typespec/http-client-java.use-eclipse-language-server=false"""
+  }
+
   $tspTrace = "--trace import-resolution --trace projection --trace http-client-java"
   $tspCommand = "npx --no-install tsp compile $tspFile $tspOptions $tspTrace"
